+++ src/main/java/spazley/scalingguis/client/gui/GeneralSettingsScreen.java
@@ -55,7 +55,7 @@ final class GeneralSettingsScreen extends Screen {
 
     @Override
     public void render(GuiGraphics graphics, int mouseX, int mouseY, float partialTick) {
-        renderBackground(graphics, mouseX, mouseY, partialTick);
+        renderBackground(graphics);
         graphics.drawCenteredString(font, title, width / 2, 15, 0xFFFFFF);
         super.render(graphics, mouseX, mouseY, partialTick);
     }