+++ src/main/java/spazley/scalingguis/client/gui/ScalingConfigScreen.java
@@ -59,7 +59,7 @@ public final class ScalingConfigScreen extends Screen {
 
     @Override
     public void render(GuiGraphics graphics, int mouseX, int mouseY, float partialTick) {
-        renderBackground(graphics, mouseX, mouseY, partialTick);
+        renderBackground(graphics);
         graphics.drawCenteredString(font, title, width / 2, 15, 0xFFFFFF);
         super.render(graphics, mouseX, mouseY, partialTick);
     }