+++ src/main/java/spazley/scalingguis/client/gui/StringSetScreen.java
@@ -109,7 +109,7 @@ final class StringSetScreen extends Screen {
 
     @Override
     public void render(GuiGraphics graphics, int mouseX, int mouseY, float partialTick) {
-        renderBackground(graphics, mouseX, mouseY, partialTick);
+        renderBackground(graphics);
         graphics.drawCenteredString(font, title, width / 2, 13, 0xFFFFFF);
         graphics.drawCenteredString(font, Component.translatable("scalingguis.config.page", page + 1, pageCount),
                 width / 2, 25, 0xA0A0A0);